requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
pydantic[email]>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
matplotlib>=3.7.0
plotly>=5.18.0
